requirements.txt
ifaddr==0.2.0
sockets==1.0.0
argparse==1.4.0
colorama==0.4.6
prettytable==3.5.0
pypcapfile==0.12.0